Add lambda_closure and Define unorthodox extended move
The unorthodox extended move just allows for easy description of
multiple moves to the same set of states. This commit also adds a
function that converts a set of these extended moves to ordinary
non-deterministic moves, in order to build an NFA.
The unorthodox moves are used to define a test NFA taken from
my lecture 4 notes from my automata class.